Added projectiles & SpinnerEnemy spawning/movement
1) Projectiles now spawn under player and travel towards mouse clicks
(unless paused) and are removed when out of bounds, BUT the projectile
speeds need to be normalized. Right now, it depends on the slope of the
line though the player position and the click position, so extreme
slopes make the projectile move too quickly.

2) SpinnerEnemies simply pick a random spot on screen to move to. There
is an limit of 10 enemies on screen for this first wave.